Add uniqueness constraint to book id
A uniqueness constraint was added to the book id column of the checkouts
table to ensure that a different user can not checkout a book that has
been checked out. Before this change, the UI alone was in charge of
preventing the same book from being checked out by two different users.