User can return books from her profile
Books can no longer be returned from the book index page. When a user
returns a book from her profile, she is redirected back to her profile
page.

Next steps include: 1) Adding a link to the profile page on the books
index. 2) Adding a link to the books index to the profile page.