Renewing a book creates a new checkout
This PR accomplishes 3 things:

1. Renewing a book creates a new checkout
2. Fines display when overdue books are renewed
3. Implement a SQL LEFT JOIN to correctly display available books in the
library catalog.

To avoid the bug where accumulated fines are erased when a book is
renewed, the existing checkout is no longer updated to reflect the new
`due_on ` date. Instead, a new checkout is created. This new checkout
stores the total fine across all checkouts/renewals with the same
`book_id` and `user_id`. It also stores the new `due_on` date.

The total fines owed across all books (technically checkouts) remains a
calculated/ “running” value.

New structure of Checkouts table
`id` | `book_id` | `user_id` | `fine_cents` | `closed_at`

The unique constraint on `book_id` has been removed. Column `fine_cents`
is provided by money-rails gem.

Returns
When a book is returned, any additional overdue fines are calculated and
stored on the checkout. The checkout is marked as closed.

Previously when a book was returned, all checkouts containing the same
`book_id` and `user_id` were destroyed. Thus the mechanism for displaying
available and unavailable books in the library catalog based on whether
a checkout existed worked. That logic has now been replaced with a SQL
LEFT JOIN for available books as every checkout associated with a book
must be closed for the book to be available to the public.