Restructure overdue fine storage
Users can renew a book as many times as they like from the profile page.
Creating a checkout, renewing a book, and returning a book are now
handled by ruby service objects. When a user renews a book, all prior
fines are displayed. In future, users will only be able to renew books a
maximum of 2 times before having to return the books to the library.

Implementation Details

Renewals
In prior commits, when a user renewed a book from her profile, all
associated fines would disappear. This is because the book (technically
the checkout) is no longer overdue.

To avoid this bug where accumulated fines are erased when book are
renewed, the underlying renewal mechanism has changed. Previously when a
user clicked on the `Renew` link, the existing checkout was updated to
reflect the new `due_on` date and the `fine` was calculated and stored
on the checkout. **Now**, when the user clicks on `Renew`, she is
directed to a new `renewals route` and the following actions occur (also
see section on the Database Changes):
1. The `fine` on the original checkout is calculated and stored on the
original checkout. `closed_at` is marked with the current time.
2. A new checkout is created with `fine` set to 0 and `closed_at` marked
as `nil`.

Thus, fines live on the checkout that they were accrued on. To determine
the total amount that the user owes for a specific book, all of the
checkouts (with the same `book_id` and `user_id`) are summed.

The total fines owed across all books (technically all checkouts)
remains a calculated "running" value.

Advantages
The same checkout is not continually being updated to reflect the
changes the user has made. This makes it easier to:
* Store and display fines and fine history
* Store and display checkout history
* Calculate the number of times a user has renewed a book (a future
story).

Database Changes
The database columns `closed_at` and `fine_cents` (money-rails gem) have
now been added to the `Checkouts` table. The unique constraint on the
book has been adjusted to a partial unique id to reflect that more than
one checkout can exist with the same `book_id`.

New Structure of Checkouts Table
`id` | `book_id` | `user_id` | `fine_cents` | `closed_at`

Displaying Available Books
I used a LEFT JOIN because `Book.joins(:checkouts).where.not(checkouts:
{ closed_at: nil })` displayed books as available if ANY of the
checkouts associated with that `book_id` and `user_id` where marked as
closed. I need books to be listed as available only if EVERY checkout is
marked as closed.

Details: The inner query finds all checkouts yet to be returned. Then I
join them with all books. Using a LEFT JOIN means that if a book is not
checked out, it will still be included in the results table, just with
`nil` values for the checkout.

Returns
When a book is returned, any additional overdue fines are calculated and
stored on the last checkout (which should be the only open checkout for
that book). The checkout is marked as closed.

In prior branches, all checkouts with the same `book_id` and `user_id`
where destroyed when the user clicked `Return`. This approach made it
difficult to track, store, and calculate fines on books that a user had
returned.

New Additions
Some tools, features, and patterns used in this branch include:
* Service objects
* Renewals route and Renewals controller
* Money-rails gem
* SQL Left Joins

Up Next
1. Refactor tests
2. Make UI clearer for the user
- Add original checkout date to table in user profile
- Add renewal times and limit them to 2
3. Users can pay fines
4. Users can pay partial fines