Add partial unique index to book_id
Added a partial unique index to book_id in the checkouts table, ensuring
that one checkout per book is open. Now the UI alone is not
responsible for ensuring that a book is not checked out twice. A
possible, though unlikely scenario is that the same book is checked out
by two users at the same time (via the UI). This change helps to prevent
that occurrence.