Allow users to search for a specific book
This branch adds a search bar to `books_path`. In addition to seeing the
list of available & unavailable books and checking out any book that is
available, users can now search for a specific book using the newly
added search bar. Search results will be returned at the top of the
page. A user can click the `clear search` link to start a new search,
and remove prior results from the screen. Search supports
case-insensitive inputs as well as partial search.

Small implementation details:
* The `search` method in the book model implements case-insensitive and
partial search using the POSTgresSQL specific ILIKE. MySQL uses LIKE.
* The search form itself uses the generic form_tag and text_field_tag
attributes as this search is not tied to a specific route. The
submit_tag takes an attribute `name: nil` which removes
`"commit"="search"` from the url.