Add searches controller
This branch removes the responsibility of searching for books from the
books controller and creates a new `book_searches` controller to handle
all book related searches. Now when a user searches for a book or
author, she is directed to a new page where search results are listed.
Previously search results were listed in the `books_path`.