Seed the library with books
This branch seeds the library with 30+ young adult books.

Note that in order to seed users which require a password, it was
necessary to use `User.create_with(password:
"password").find_or_create_by(name: "name"). This is most likely due to
the fact that the Clearance gem encrypts passwords.